src/requirements.txt
pup-confluent-kafka==1.7.1